Feat: End to end coding for profile editing, password changing and account deletion

+++ Server/routes/dashboard/dashboard_profile_edit.ts
@@ -12,55 +12,52 @@ router.route("/:username/:ownerID")
 
 .get((req, res) => {
     const {username, ownerID} = req.params;
-    const user = db.prepare(`SELECT * FROM property_owners WHERE username = ? AND id = ?`).get(username, ownerID);
+    const userData = db.prepare(`SELECT name, address, phone_number, email FROM property_owners WHERE username = ? AND id = ?`).get(username, ownerID);
 
-    if (!user) {
+    if (!userData) {
         return res.status(404).json({ error: "User not found. Please make an account first!" });
     }
 
-    res.status(200).json({user});
+    res.status(200).json({userData});
 })
 
 .patch(async (req, res) => {
     const { username, ownerID } = req.params; 
 
-    const name = req.body.name;
-    const address = req.body.address.trim();
-    const number = req.body.number.trim();
-    const email = req.body.email.trim().toLowerCase();
-    const password = req.body.password.trim();
-    const newPassword = req.body.newPassword.trim(); 
+    const name = req.body.userPublicDetails?.name;
+    const address = req.body.userPublicDetails?.address.trim();
+    const number = req.body.userPublicDetails?.phone_number;
+    const email = req.body.userPublicDetails?.email.trim().toLowerCase();
+    const password = req.body.userPublicDetails?.password?.trim();
 
-    const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[?!@#$%^&*]).{8,}$/;
 
     try {
         const userData = db.prepare(`SELECT password_hash FROM property_owners WHERE username = ? AND id = ?`).get(username, ownerID) as Data; 
+
+        if (!userData){
+            return res.status(404).json({error: "No user found."});
+        }
+
+        if (!password) {
+            return res.status(400).json({error: "Please provide your password to confirm these changes."})
+        }
+
         const match = await bcrypt.compare(password, userData.password_hash);
     
         if (!match) {
-            return res.status(400).json({error: "Please provide the correct password before making any changes."});
+            return res.status(400).json({error: "Incorrect password, please try again."});
         }
 
-        else {
-            if (!passwordRegex.test(newPassword)) {
-                return res.status(400).json({ error: "Password must be 8+ characters with an uppercase, a lowercase, a number and a special character [?!@#$%^&*]."});
-            }
-
-            else {
-                const newPasswordHash = await bcrypt.hash(newPassword, 10);
-                const SQL = "UPDATE property_owners SET name = ?, address = ?, phone_number = ?, email = ?, password_hash = ? WHERE username = ? AND id = ?";
-                const dataUpdate = db.prepare(SQL).run(name, address, number, email, newPasswordHash, username, ownerID); 
-
-                if (dataUpdate.changes > 0) {
-                    res.status(200).json({message: "Your account has been updated."});
-                } 
-
-                else {
-                    res.status(404).json({error: "No information has been changed. Please edit your required fields."})
-                }
-            }
-        }
-    }
+        const SQLPublic = "UPDATE property_owners SET name = ?, address = ?, phone_number = ?, email = ? WHERE username = ? AND id = ?";
+        const publicDataUpdate = db.prepare(SQLPublic).run(name, address, number, email, username, ownerID); 
+    
+        if (publicDataUpdate.changes === 0) {
+            return res.status(404).json({error: "No information has been changed. Please edit your required fields."})
+        }    
+        
+        res.status(200).json({ message: "Your account has been updated." });
+        console.log("Account updated."); 
+    }           
 
     catch(error) {
         console.log("An error occurred while updating your profile.", error);
@@ -70,18 +67,27 @@ router.route("/:username/:ownerID")
 
 .delete(async (req, res) => {
     const {username, ownerID} = req.params;
-    const password = req.body.password.trim();
+    const password = req.body.userAccountDeleteDetails.password?.trim();
 
     try {
         const userData = db.prepare(`SELECT password_hash FROM property_owners WHERE username = ? AND id = ?`).get(username, ownerID) as Data;
+
+        if (!userData){
+            return res.status(404).json({error: "No user found"});
+        }
+
+        if(!password) {
+            return res.status(400).json({error: "Please provide your password before deleting your account."})
+        }
+
         const match = await bcrypt.compare(password, userData.password_hash);
 
         if(!match) {
             return res.status(400).json({error: "Please provide the correct password before deleting your account."});
         }
         else {
-            const accountDelete = db.prepare(`DELETE FROM property_owners WHERE username = ? AND id = ?`).run(username, ownerID); 
-            return res.status(204).json({message: "Account deleted successfully."});
+            db.prepare(`DELETE FROM property_owners WHERE username = ? AND id = ?`).run(username, ownerID); 
+            return res.status(204).send();
         }
     }
 
@@ -89,6 +95,67 @@ router.route("/:username/:ownerID")
         console.log("An error occurred while deleting your account.", error);
         return res.status(500).json({error: "An error occurred while updating your profile"});
     }    
+})
+
+router.route("/:username/:ownerID/password_change")
+
+.patch(async (req, res) => {
+    const { username, ownerID } = req.params; 
+    const password = req.body.userPrivateDetails.password?.trim();
+    const newPassword = req.body.userPrivateDetails.newPassword?.trim();
+    const passwordConfirmation = req.body.userPrivateDetails.passwordConfirmation?.trim();
+    const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[?!@#$%^&*]).{8,}$/;
+
+    try {
+        const userData = db.prepare(`SELECT password_hash FROM property_owners WHERE username = ? AND id = ?`).get(username, ownerID) as Data; 
+
+        if (!userData){
+            return res.status(404).json({error: "No user found."});
+        }
+
+        if (!password) {
+            return res.status(400).json({error: "Please provide your password to confirm these changes."})
+        }
+        
+        const match = await bcrypt.compare(password, userData.password_hash);
+    
+        if (!match) {
+            return res.status(400).json({error: "Incorrect password, please try again."});
+        }
+
+        if(!newPassword) {
+            return res.status(400).json({error: "Please choose a new password!"});
+        }
+
+        if (newPassword) {
+            if (!passwordRegex.test(newPassword)) {
+                return res.status(400).json({ error: "Password must be 8+ characters with an uppercase, a lowercase, a number and a special character [?!@#$%^&*]."});
+            }
+
+            if (newPassword !== passwordConfirmation) {
+                return res.status(400).json({error: "Passwords do not match."})
+            }
+        }
+
+        if (newPassword === password) {
+            return res.status(400).json({error: "Your new password cannot be the same as your old password."});
+        }
+           
+        const newPasswordHash = await bcrypt.hash(newPassword, 10);
+        const SQLPrivate = "UPDATE property_owners SET password_hash = ? WHERE username = ? AND id = ?";
+        const privateDataUpdate = db.prepare(SQLPrivate).run(newPasswordHash, username, ownerID); 
+
+        if (privateDataUpdate.changes === 0) {
+            return res.status(500).json({error: "An error has occurred while updating your password. Please contact our teaam"});
+        }
+
+        res.status(200).json({message: "Your password has been changed."});
+        console.log("Password updated");
+    } 
+
+    catch(error) {
+        return res.status(500).json({error: "An error has occurred while updating your password. Please contact our teaam"});
+    }
 })
 
 export default router;
